Resolves: fdo#34987 skip autoheight reset if it will be set to the same value
Triggered by 792561f2db03f31e3af906a3ee95f87e943f095a which explicitly sets
rSet.Put( SdrTextAutoGrowHeightItem(FALSE) ); so there is something set on the
style which is being overwritten.

The code here resets the style to the default of "true" before going on to set
it to the explicit "false" again. In that window of time the master shapes
listen to the property change, on being set to autoheight they resize and on
being unset, they remain stuck on their autoheight calculated size.

Change-Id: I567a791b2bbbcb3a1a111633fabf509142984645
Reviewed-on: https://gerrit.libreoffice.org/5887
Reviewed-by: mhofmann <borim7@web.de>
Reviewed-by: Thorsten Behrens <thb@documentfoundation.org>
Tested-by: Thorsten Behrens <thb@documentfoundation.org>