Bin outdated and overly specific "requirements" for OS X and Windows
Surely we list better requirements on a website anyway. The ones
listed here were either plainly wrong or of historical interest
only. These "requirements" have no legal meaning anyway, so surely we
can just assume common sense instead of trying to figure out if we
"require" 256 megabytes of RAM or 512.

Somebody please have the courage to edit the Linux section, too. It is
woefully out of date, over-specific and misleading.

Change-Id: Id9c7208538e55ece5a16dae2c120e77dd7839b04