Fix drawing of fine dashed border type
On the Borders tab in the preview block (called User-defined area
in help) fine dashed border is drawn as solid line.
The problem is that 1 pixel gap is too small and painting
can hadnle it so I doubled the numbers to get the same rate and
get the rigth visual result.

Change-Id: Ide037fa3f4f90b304dc53f5f806c2acd2194ec0c