installer: put in a hack for MacOSX to find language pack files
The files are in LibreOffice.app, are packaged in LibreOffice Language
Pack.app and will then be copied into LibreOffice.app when the lang-pack
is clicked by the user.

This also neatly avoids having to deal with file names with spaces in
make.

Also, root directory is $(PRODUCTNAME).app which works with
--disable-release-build too.

Change-Id: Ie064831316d790b137022a1632721c62500a46a1