Theme icons are loaded only once
With the theme icon cleaning commit
0dc3f1db1bf8ab8fea1f53d3675ea7fbb63f5aeb, duplicated
icons (same icon for multiple actions) doesn't need to
be stored in different files. The file links.txt inside
the theme pack can handle filename redirections.
So, it keeps the theme pack smaller. However, the
duplicates keep being loaded, because they use different
names.

With this patch, the names of those duplicates are mapped
also when querying for the icons. This way, an icon is
loaded only once, saving RAM.

Change-Id: Ieb5b8424c1515f0033276bf314cdef02a4e01862
Signed-off-by: Rodolfo Ribeiro Gomes <rodolforg@gmail.com>
Reviewed-on: https://gerrit.libreoffice.org/5943
Reviewed-by: Caolán McNamara <caolanm@redhat.com>
Tested-by: Caolán McNamara <caolanm@redhat.com>