Fixes for cross-compilation on OS X (to iOS)
Introduce SDKDIRNAME as a configury variable and use it instead of the
gbuild gb_Package_SDKDIRNAME. Then we can easily construct the
SDKDIRNAME_FOR_BUILD variant that is needed to find the specially
named SDK in instdir on OS X when cross-compiling.

Move the version number section in configure.ac earlier.

Change-Id: Iee3db1a50ad4c7a9f91bbc5e0d0b01d76a76f701