merge queue_layout and queue_resize
so that any window derived class, and not just dialogs, can trigger layouting
of their children. Merge together the handful of hacked-up impls of this.

Do that then for the sidebar PanelLayout so that when the label of the custom
animation frame changes that the frame allocates enough space for the new label
to display fully

Change-Id: I9a95f6c3f60cd6cea47656e66cb9ffcc154a3a5a