make dev-install is obsolete
There is no need for developers to run it as part of the daily workflow
because instdir is much easier to use, so just let it print an error
message.  Also remove the "install" symlink.

Provide the previous dev-install functionality as "test-install" target
so that it is possible to quickly create some kind of actual
installation against which the subsequentchecks can run by
setting OOO_TEST_SOFFICE, to e.g. check that scp2 isn't missing files.

Change-Id: I8c0b05a87f55b1db8c1d1e76f64ab64b9cc695eb