StyleSheetTable::ApplyStyleSheets: stop clearing NumberingStyleName manually
This was introduced in commit 9c7ffb6b61f7c94ca5b38d11be1807f1e29ff4bc
([cbosdo03]Applied the patches to upstream, 2009-12-02), as setting the
outline level also set the NumberingStyleName to "Outline" as well,
which is not what we want.

Since then, commit 3c4a2375c41398e01ff856d8ad96f8d70a918079 (Don't load
the default values of the styles in writerfilter, 2012-08-20) already
disabled default styles, so the previous manual clearing is no longer
necessary (testcases already fail if this later commit is reverted).

As a result, just remove this manual clearing, it's no longer necessary
and it is problematic, as it triggers copying the SvxLRSpaceItem from
the parent style with no good reason.

Change-Id: Ie17547a979a6b0feadbd2eed0f06e84fcb17fca6