fdo#36868 WW8 import: allow outline numbering and list style in the same ...
.. paragraph style

The original problem (from a user's point of view) was that the second
level of the numbering started from 1.1 instead of 2.1 in the bugdoc.
This was fixed by using outline numbering for level 2 as well, but this
is problematic in many cases: we want to have outline numbering exactly
when outline numbering is enabled for the given paragraph style.

So revert the change in SwWW8ImplReader::SetStylesList() and fix it
differently: SwWW8ImplReader::RegisterNumFmtOnStyle() explicitly ignores
list style if outline numbering is available with no good reason. Both
the WW8 format and Writer core allows to have outline numbering and a
list style at the same time, so set list style even when outline
numbering is available. This fixes the original issue, too -- without
introducing nasty fake outline numbering usage.

Also add a testcase for the original issue.

(regression from e6eeeb61bbdbd9cd5020808b3da5017d8b310bae)

Change-Id: Id7d2d67a96a858aee3230110cb518fea51d19d38